requirements.txt
Babel
python-gettext
PyInstaller
mypy
black
flake8
pygame
Pillow
fs
